+++ final/FinalProject/Imperium.cs
@@ -1,7 +1,7 @@
 public abstract class Imperium : Faction
 {
-    public Imperium(string name, string tagline, string description, string modelName, int modelCost) : base(name, tagline, description, modelName, modelCost)
+    public Imperium(string name, string tagline, string description, string modelName, double modelCost) : base(name, tagline, description, modelName, modelCost)
     {
-        
+
     }
 }
